Remove set_up_django.sh and its usage in generator/index.js
I figure since this is a js project it's better to do things in node than in a bash script.
It also works a little nicer for the end user since it outputs the messages onto the same command line they are using instead of in a pop up shell.